fix rename toolbar
